BugFix: error in geometry culling calculation.
git-svn-id: svn://svn.code.sf.net/p/glmixer/Source/trunk@1713 a4deb95c-d308-4ec4-8dc1-6419bbeeb61a

+++ src/Source.cpp
@@ -547,8 +547,8 @@ GLuint Source::getTextureIndex() const {
 
 void Source::testGeometryCulling() {
 
-    int w = SOURCE_UNIT;
-    int h = SOURCE_UNIT;
+    double w = SOURCE_UNIT;
+    double h = SOURCE_UNIT;
 
     if (OutputRenderWindow::getInstance()->getAspectRatio() > 1)
         w *= OutputRenderWindow::getInstance()->getAspectRatio();
@@ -561,7 +561,7 @@ void Source::testGeometryCulling() {
     else {
         // not obviously visible
         // but it might still be parly visible if the distance from the center to the borders is less than the width
-        int d = sqrt( scalex*scalex + scaley*scaley);
+        double d = sqrt( scalex*scalex + scaley*scaley);
         if ((x + d < -w) || (x - d > w))
             culled = true;
         else if ((y + d < -h) || (y - d > h))
